Tests: binary trie E2E + proofs + spec vectors + SSZ round-trip
- BinaryDevChainE2ETests: embedded binary DevChain with jsign-
  validated genesis root, block production, proof verification
- BinaryProofServiceTests: account + storage proof generation and
  cryptographic verification against state root
- BinaryIncrementalStateRootTests: incremental vs full state root
  equivalence, warm-start from stored root
- BinaryTrieSpecVectorTests: extended jsign cross-validation vectors
- RocksDB tests: binary trie node store, storage, code_hash layout,
  three-way comparison (InMemory vs RocksDB vs SQLite)
- AccountLayoutStrategyTests: Patricia vs Binary layout round-trip
- BlockEncodingProviderRoundTripTests: SSZ encode/decode parity
- Fixed: TraceConverterCallTraceTests (CallInput→EvmCallContext),
  HistoricalStateStoreTests (BigInteger vs EvmUInt256 Assert.Equal)

+++ tests/Nethereum.CoreChain.UnitTests/Storage/HistoricalStateStoreTests.cs
@@ -266,7 +266,7 @@ namespace Nethereum.CoreChain.UnitTests.Storage
             var newest = await diffStore.GetNewestDiffBlockAsync();
             var atNewest = await store.GetAccountAtBlockAsync(Addr1, newest.Value);
             Assert.NotNull(atNewest);
-            Assert.Equal(100 + newest.Value * 10, atNewest.Balance);
+            Assert.Equal(100 + newest.Value * 10, (BigInteger)atNewest.Balance);
         }
 
         [Fact]
